Move functions to specific files
